+++ fiddles/bash/fiddle-0051-InstallCucumber/script.sh
@@ -1,0 +1,18 @@
+#!/usr/bin/env bash
+
+clear;
+echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
+echo "Bash version ${BASH_VERSION}..."
+
+function installIstanbul() {
+    npm install -g istanbul;
+}
+
+function installCucumber() {
+    npm install -g cucumber;
+
+}
+installIstanbul;
+installCucumber;
+
+