+++ fiddles/bash/template/test.sh
@@ -3,30 +3,15 @@ clear;
 echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
 echo "Bash version ${BASH_VERSION}..."
 
-test1=$(./script.sh "svg" "0022";);
-test2=$(./script.sh "svg" "22";);
-test3=$(./script.sh "svg" "222222";);
+test1=$(./script.sh 1;);
 
 
-if [[ ! ${test1} -eq "fiddle-0022-Spirograph" ]]
+if [[ ! ${test1} -eq "2" ]]
 then
     echo "test 1 failed"
 else
     echo "test 1 passed"
 fi
 
-if [[ ! ${test2} -eq "fiddle-0022-Spirograph" ]]
-then
-    echo "test 2 failed"
-else
-    echo "test 2 passed"
-fi
-
-if [[ ! ${test3} -eq "Invalid Criteria specified!" ]]
-then
-    echo "test 3 failed"
-else
-    echo "test 3 passed"
-fi