+++ fiddles/chrome/fiddle-0003-BookmarkTable/app.js
@@ -1,8 +1,0 @@
-(function (app, $, undefined) {
-  $(document).ready(function () {
-        // fiddle
-        var hook = $('#fiddleHook');
-        hook.html('<div class="text-info"><i> Hi There! </i></div>');
-    });
-})(window.app = window.app || {}, jQuery)
-