+++ fiddles/python/fiddle-0001-BeautifulSoup/script.py
@@ -1,0 +1,23 @@
+from urllib.request import urlopen
+from urllib.error import HTTPError
+from urllib.error import URLError
+from bs4 import BeautifulSoup
+
+def getTitle(url):
+    try:
+        html = urlopen(url)
+    except HTTPError as e:
+        return None
+    try:
+        bsObj = BeautifulSoup(html.read(), "html.parser")
+        title = bsObj.body.h1
+    except AttributeError as e:
+        return None
+    return title
+
+title = getTitle("http://three-bradyhouse.rhcloud.com/")
+
+if title == None:
+    print("Title could not be found")
+else:
+    print(title)