+++ scripts/bin/tpl/jasmine_stub_header.js
@@ -1,2 +1,2 @@
-/*jslint browser: true, vars: true, plusplus: true, devel: true, nomen: true, indent: 4, maxerr: 50, todo: true, sloppy: true, unparam: true */
-/*global Ext, window */
+/* jshint unused: vars */
+/* global Ext, Risk */